pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "whoop-write-mcp"
version = "0.1.0"
description = "MCP server for Whoop with write support — log activities, set alarms, update weight, read biometrics"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [{name = "jd1207"}]
dependencies = [
    "mcp[cli]>=1.0",
    "whoop-write-api>=0.4.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
whoop-write-mcp = "whoop_write_mcp.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
